Create OpenCVTest.py
This is the code I have been using to play around with the OpenCV library. You will need to pip install opencv, cvzone, and mediapipe to use it. Also you will need to go into cvzone and change findDistance so that it unpacks z1 and z2 or else it will kick back an error.